refactor: adContentMap 파일 통일
