--- src/shared/widget/Ad/AdContent.ts
@@ -1,14 +1,0 @@
-export const adContentMap = {
-  callenge: {
-    image: '/images/ad/img_ad_1.png',
-    title: '캘린지 1주년! 생일 축하해 🎂',
-    caption: '영화 6편을 골라보고 캘린지 완주해 보세요',
-    bottomCaption: '왓챠피디아',
-  },
-  conclave: {
-    image: '/images/ad/img_ad_2.png',
-    title: '선택을 기다리는 고요한 전장, 콘클라베',
-    caption: '~5/1(목)까지 20% 할인가로 만나보세요!',
-    bottomCaption: '왓챠 개별 구매',
-  },
-} as const
